maint #342 extract ophyd device helpers from misc.py into devices.py
Create hklpy2/devices.py with 7 ophyd device construction helpers:
VirtualPositionerBase, define_real_axis, dict_device_factory,
dynamic_import, make_component, make_dynamic_instance, parse_factory_axes.

Remove all 7 definitions from misc.py (no re-export shim). Update
internal imports in diffract.py (deferred imports inside creator()) and
test files. dev_i182.py is gitignored and not tracked.

misc.py now imports nothing from ophyd directly except Device
(still needed by ConfigurationRunWrapper).

Agent: OpenCode (claudesonnet46)
