+++ src/hklpy2/tests/test_misc.py
@@ -24,24 +24,24 @@ from yaml.parser import ParserError
 
 from ..diffract import creator
 from ..diffract import diffractometer_class_factory
-from ..misc import ConfigurationRunWrapper
+from ..devices import VirtualPositionerBase
+from ..devices import dict_device_factory
+from ..devices import dynamic_import
+from ..devices import make_dynamic_instance
+from ..devices import parse_factory_axes
 from ..exceptions import NoForwardSolutions
 from ..exceptions import SolverError
-from ..misc import VirtualPositionerBase
+from ..misc import ConfigurationRunWrapper
 from ..misc import axes_to_dict
 from ..misc import compare_float_dicts
 from ..misc import convert_units
-from ..misc import dict_device_factory
 from ..misc import distance_between_pos_tuples
-from ..misc import dynamic_import
 from ..misc import flatten_lists
 from ..misc import get_run_orientation
 from ..misc import get_solver
 from ..misc import istype
 from ..misc import list_orientation_runs
 from ..misc import load_yaml_file
-from ..misc import make_dynamic_instance
-from ..misc import parse_factory_axes
 from ..misc import pick_closest_solution
 from ..misc import pick_first_solution
 from ..misc import roundoff
